+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# 0.2.0
+
+## all services
+
+* Upgrade Dependencies
+* Migrate `chrono` to `time`
+
+## bcr-wdc-relay
+
+* Add `relay_url` config parameter (ENV: `RELAY_URL`), which must be set to the public facing URL of the relay
+
 # 0.1.0
 
 ## bcr-wdc-eic-service